% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecfg.R
\name{BaseCfg}
\alias{BaseCfg}
\title{BaseCfg}
\description{
A base R6 class that supports building declarative config
  classes which auto-populate from internal defaults, environment variables,
  and command line arguments.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseCfg-new}{\code{BaseCfg$new()}}
\item \href{#method-BaseCfg-clone}{\code{BaseCfg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseCfg-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseCfg-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseCfg$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseCfg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseCfg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseCfg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
