% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{Create a new `Opt` instance}
\usage{
opt(...)
}
\arguments{
\item{...}{The parameters to be passed to the `Opt` constructor.}
}
\value{
A new `Opt` instance.
}
\description{
This function creates a new `Opt` instance with the provided parameters.
}
